requirements.txt
openai==1.59.6
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0
python-dotenv==1.0.1
fastapi[standard]==0.115.6
uvicorn==0.34.0